# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API for managing tasks, now with Supabase authentication and billing."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },

    "/auth/login": {
      "post": {
        "summary": "Log in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },

    "/auth/logout": {
      "post": {
        "summary": "Logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out"
          }
        }
      }
    },

    "/public/info": {
      "get": {
        "summary": "Public endpoint",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },

    "/protected/profile": {
      "get": {
        "summary": "User profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile"
          }
        }
      }
    },

    "/protected/dashboard": {
      "get": {
        "summary": "Dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard"
          }
        }
      }
    },

    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "Task list"
          }
        }
      },
      "post": {
        "summary": "Create task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          }
        }
      }
    },

    "/tasks/{id}": {
      "get": {
        "summary": "Get task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },

      "put": {
        "summary": "Update task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "done": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          }
        }
      },

      "delete": {
        "summary": "Delete task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          }
        }
      }
    },

    "/usage": {
      "post": {
        "summary": "Record billable usage",
        "description": "Records API calls or AI token usage with idempotency protection and quota enforcement.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tenant_id",
                  "type",
                  "quantity",
                  "idempotency_key"
                ],
                "properties": {
                  "tenant_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "api_call",
                      "ai_token"
                    ]
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 1
                  },
                  "idempotency_key": {
                    "type": "string",
                    "example": "request-123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usage recorded"
          },
          "200": {
            "description": "Duplicate ignored"
          },
          "404": {
            "description": "Tenant not found"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    }
  }
}